Publish to crates with our own action
It seems the action we've been using updates `Cargo.lock`. This is
problematic, because our tests run against some particular `Cargo.lock`,
and this makes it so that the published version cannot be installed with
the exact same dependencies (using `install --locked`).

In another PR I will update the deps and fix whatever the actual problem
is. I think we could also consider dependabot to keep dependencies up to
date.