Fix incorrect codegen on event fields with generic types
On the generation of code for an event with a field that mixed:

  1. Non-user defined generic (e.g. Option<T>)
  2. User-defined generic parameter (e.g. T = CustomStruct)

The generation failed to prefix the user-defined parameter with "super::"
due to the generic itself not being user-defined.

An example of this issue would be `field: Option<CustomStruct>`
translating to `field: Option<CustomStruct>` instead of
`field: Option<super::CustomStruct>`.

The issue resided in the code assuming that if the generic was non-user
defined, then its parameters were also non-user defined, which as proved
above may not be the case.